splitfl_kd_demo/requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.24
tqdm>=4.66
PyYAML>=6.0
matplotlib>=3.8
